voice-agents/livekit/pyproject.toml
[project]
name = "livekit-voice-agent"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "livekit-agents[silero,turn-detector,deepgram,openai]~=1.3",
    "livekit-plugins-noise-cancellation~=0.2",
    "python-dotenv>=1.2.1",
    "flask>=3.0.0",
    "livekit-api>=0.7.0",
    "opentelemetry-sdk>=1.20.0",
    "requests>=2.31.0",
]
